[desktop] TransfersTab is notified of transfer selection events
BTDownloadMediator is the Component Mediator that holds the transfer's table, it now has a new listener that notifies of a BTDownload object being selected. If nothing is selected it fires up a callback with a null selection value.

This listener is implemented by TransfersTab which decides what to do upon the given selection.
If FrostWire is running with a non-split Search/Transfers view and we've been given a Torrent download then the TransfersTab should display the TransferDetails component, otherwise it should hide it.

Next up, the TransferDetails component creation